requirements.txt
pandas>=2.2
numpy>=1.26
yfinance>=0.2.38
PyYAML>=6.0.2
scipy>=1.13
matplotlib>=3.8
pytest>=8.3
